Refactor: Backend architecture is created
