--- backend/controllers/Test.go
@@ -1,10 +1,0 @@
-package controllers
-
-import(
-	"github.com/gin-gonic/gin"
-	"net/http"
-)
-
-func Test(c *gin.Context) {
-    c.String(http.StatusOK, "Hello World!")
-}
