+++ backend/controllers/health_check.go
@@ -1,0 +1,11 @@
+package controllers
+
+import (
+	"net/http"
+
+	"github.com/gin-gonic/gin"
+)
+
+func HealthCheck(c *gin.Context) {
+	c.String(http.StatusOK, "Hello World!")
+}
