+++ backend/main.go
@@ -1,10 +1,30 @@
 package main
 
 import (
-	"Syojincoder2/router"
+	"database/sql"
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/dot_p/syojin/api"
 )
 
-func main(){
-	r := router.SetupRouter()
+var (
+	dbUser     = os.Getenv("DB_USER")
+	dbPassword = os.Getenv("DB_PASSWORD")
+	dbDatabase = os.Getenv("DB_NAME")
+	dbConn     = fmt.Sprintf("%s:%s@tcp(127.0.0.1:3306)/%s?parseTime=true", dbUser, dbPassword, dbDatabase)
+)
+
+func main() {
+	db, err := sql.Open("mysql", dbConn)
+	if err != nil {
+		log.Println("Fail to connect DB")
+		return
+	}
+
+	r := api.NewRouter(db)
+
+	log.Println("server start at port 8080")
 	r.Run(":8080")
 }
