--- backend/models/User.go
@@ -1,5 +1,0 @@
-package models
-
-type User struct {
-    UserName string `json:"userName"`
-}
